+++ src/shared/components/results/ExactFieldSuggestion.tsx
@@ -1,31 +1,15 @@
-import { exactMatchSearchTerms, SearchTextLink } from './SearchSuggestions';
-
 import {
-  parse,
-  stringify,
-} from '../../../query-builder/utils/queryStringProcessor';
-
-import { Clause } from '../../../query-builder/types/searchTypes';
+  exactMatchSearchTerms,
+  modifyQueryWithSuggestions,
+  SearchTextLink,
+} from './SearchSuggestions';
 
 const ExactFieldSuggestion = ({ query }: { query: string }) => {
-  let modifiedClauses: Clause[] = [];
-  let searchValue = '';
-
-  const parsedQuery = parse(query);
-  modifiedClauses = parsedQuery.map((clause) => {
-    if (exactMatchSearchTerms.includes(clause.searchTerm.term)) {
-      const queryBit = clause.queryBits;
-      const modifiedQueryBit: Record<string, string> = {};
-      Object.entries(queryBit).forEach(([k, v]) => {
-        modifiedQueryBit[`${k}_exact`] = v;
-        searchValue = v;
-      });
-      return { ...clause, queryBits: modifiedQueryBit };
-    }
-    return { ...clause };
-  });
-
-  const modifiedQuery = stringify(modifiedClauses);
+  const { modifiedQuery, searchValue } = modifyQueryWithSuggestions(
+    query,
+    'exact',
+    exactMatchSearchTerms
+  );
 
   if (query !== modifiedQuery) {
     return (
